// src/app/interfaces/filee.ts
export interface Filee {
    id: number;                    
    name: string;                  
    path: string;                   
    folderId: number;               
    contentType: string;            
    size: number;                   
    type: string;                   
    uploadDate: string;           
    version: number;               
    isActive: boolean;             
  }
  